getObbAssetFileDescriptorJNI() 함수에서 long 타입 변수에 long long 값을 대입하여 메모리 침범이 일어날 수 있었던 문제 수정
git-svn-id: svn://35.216.44.119/DVM/dragonvillage/src/frameworks@88300 5003bd00-c432-460b-831f-0659da9acde2

+++ cocos2d-x/cocos/platform/android/jni/Java_org_cocos2dx_lib_Cocos2dxHelper.cpp
@@ -424,8 +424,8 @@ int getObbAssetFileDescriptorJNI(const char* path, long* startOffset, long* size
             jboolean copy = JNI_FALSE;
             jlong *array = methodInfo.env->GetLongArrayElements(newArray, &copy);
             fd = static_cast<int>(array[0]);
-            *startOffset = array[1];
-            *size = array[2];
+            *startOffset = static_cast<long>array[1];
+            *size = static_cast<long>array[2];
             methodInfo.env->ReleaseLongArrayElements(newArray, array, 0);
         }
